requirements.txt
coinbase-advanced-py == 1.8.2
requests >= 2.31.0
urllib3 >= 2.2.2
PyYAML >= 6.0.1
cryptography >= 42.0.4
cffi
alphasquared-py >= 0.3.0
fear-and-greed-crypto >= 0.1.0
schedule >= 1.2.0
python-dotenv >= 1.0.0